+++ spec/rubocop/config_obsoletion_spec.rb
@@ -69,6 +69,7 @@ RSpec.describe RuboCop::ConfigObsoletion do
           'Layout/SpaceAfterControlKeyword' => { 'Enabled': true },
           'Layout/SpaceBeforeModifierKeyword' => { 'Enabled': true },
           'Lint/InvalidCharacterLiteral' => { 'Enabled': true },
+          'Style/MethodMissingSuper' => { 'Enabled': true },
           'Lint/RescueWithoutErrorClass' => { 'Enabled': true },
           'Lint/SpaceBeforeFirstArg' => { 'Enabled': true },
           'Style/SpaceAfterControlKeyword' => { 'Enabled': true },
@@ -195,6 +196,8 @@ RSpec.describe RuboCop::ConfigObsoletion do
           The `Lint/InvalidCharacterLiteral` cop has been removed since it was never being actually triggered.
           (obsolete configuration found in example/.rubocop.yml, please update it)
           The `Lint/SpaceBeforeFirstArg` cop has been removed since it was a duplicate of `Layout/SpaceBeforeFirstArg`. Please use `Layout/SpaceBeforeFirstArg` instead.
+          (obsolete configuration found in example/.rubocop.yml, please update it)
+          The `Style/MethodMissingSuper` cop has been removed since it has been superseded by `Lint/MissingSuper`. Please use `Lint/MissingSuper` instead.
           (obsolete configuration found in example/.rubocop.yml, please update it)
           The `Style/MethodMissing` cop has been split into `Style/MethodMissingSuper` and `Style/MissingRespondToMissing`.
           (obsolete configuration found in example/.rubocop.yml, please update it)